Breaking fields out into it's own module. Fields.py is bound to blow up and was already over 1000 lines with SecureField plus docs
