+++ cincoconfig/__init__.py
@@ -8,11 +8,12 @@
 # Public API
 from .config import Config, Schema
 from .abc import Field, AnyField
-from .fields import *
 from .formats import FormatRegistry
+from .fields import *
 
 
 __all__ = ('Config', 'Schema', 'FormatRegistry', 'StringField', 'IntField', 'FloatField',
            'PortField', 'IPv4AddressField', 'IPv4NetworkField', 'FilenameField', 'BoolField',
            'UrlField', 'ListField', 'HostnameField', 'DictField', 'VirtualField',
-           'ApplicationModeField', 'LogLevelField', 'Field', 'AnyField', 'SecureField')
+           'ApplicationModeField', 'LogLevelField', 'Field', 'AnyField', 'SecureField',
+           'NumberField')
